// frontend/src/app/core/services/api/model/qualificationSchemeDto.ts
/**
 * BFG API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * Data Transfer Object for qualification scheme
 */
export interface QualificationSchemeDto { 
    uuid?: string;
    name?: string;
    laneCount?: number;
    isActive?: boolean;
    createdAt?: string;
    modifiedAt?: string;
}