// frontend/src/app/core/services/api/model/qualificationTierRequest.ts
/**
 * BFG API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * Request payload for creating or updating a qualification tier
 */
export interface QualificationTierRequest { 
    qualificationSchemeId: string;
    boatCountMin: number;
    boatCountMax: number;
    heatCount: number;
    semiFinalCount: number;
    finalBCount: number;
    finalACount: number;
}